new_gen/backend/requirements.txt
fastapi
uvicorn
pydantic
email-validator
pillow
sqlalchemy
asyncpg
cryptography
